requirements.txt
pytorch-lightning>=2.0.0
torch-ema>=0.3.0

numpy>=1.23
scipy>=1.10
librosa>=0.9.0
soundfile>=0.12.0
einops>=0.6.0
torchfcpe>=0.0.4

tqdm>=4.64.0
matplotlib>=3.7.0
pydub>=0.25.0
pesq>=0.0.4
pystoi>=0.3.3

pyyaml>=6.0
